BARK-57: Move opListVetAccountsByVet into VetAccountService as getVetAccountsByVetId
