+++ src/lib/admin/rpc/commands/list-vet-accounts-by-vet.ts
@@ -1,12 +1,11 @@
-import { opListVetAccountsByVet } from "@/lib/bark/operations/op-list-vet-accounts-by-vet";
 import { AdminCommand, AdminCommandArgs } from "../admin-command";
 
 export class ListVetAccountsByVet implements AdminCommand {
   async run(args: AdminCommandArgs): Promise<string> {
-    const { context, request } = args;
+    const { actor, request } = args;
     const obj = JSON.parse(request);
     const { vetId } = obj;
-    const { result, error } = await opListVetAccountsByVet(context, { vetId });
+    const { result, error } = await actor.getVetAccountsByVetId({ vetId });
     if (error !== undefined) {
       throw new Error(error);
     }
